+++ src/app/search/FilterHelp.tsx
@@ -31,7 +31,8 @@ function FilterHelp({ destinyVersion }: { destinyVersion: 1 | 2 }) {
             </tr>
             <tr>
               <td>
-                <span>perk</span> <span>&quot;item name&quot;</span> <span>perk:magazine</span>
+                <span>perk</span> <span>&quot;item name&quot;</span> <span>perk:magazine</span>{' '}
+                <span>perkname:enhanced</span>
               </td>
               <td>{t('Filter.PartialMatch')}</td>
             </tr>
@@ -169,7 +170,6 @@ function FilterHelp({ destinyVersion }: { destinyVersion: 1 | 2 }) {
             <tr>
               <td>
                 <span>is:wishlist</span>
-                <span>is:curated</span>
               </td>
               <td>{t('Filter.Wishlist')}</td>
             </tr>
@@ -577,7 +577,7 @@ function FilterHelp({ destinyVersion }: { destinyVersion: 1 | 2 }) {
                   <span>source:drifter,source:gambit</span> <span>source:ep</span>
                   <span>source:raid</span> <span>source:leviathan</span> <span>source:sos</span>
                   <span>source:eow</span> <span>source:lastwish</span> <span>source:scourge</span>
-                  <span>source:crownofsorrows</span>
+                  <span>source:crownofsorrow</span>
                   <span>source:prestige</span> <span>source:adventure</span>{' '}
                   <span>source:blackarmory</span> <span>source:gambitprime</span>
                   <span>source:menagerie</span>
@@ -611,7 +611,7 @@ function FilterHelp({ destinyVersion }: { destinyVersion: 1 | 2 }) {
                   <span dangerouslySetInnerHTML={{ __html: t('Filter.Source.EoW') }} />
                   <span dangerouslySetInnerHTML={{ __html: t('Filter.Source.LastWish') }} />
                   <span dangerouslySetInnerHTML={{ __html: t('Filter.Source.Scourge') }} />
-                  <span dangerouslySetInnerHTML={{ __html: t('Filter.Source.CrownOfSorrows') }} />
+                  <span dangerouslySetInnerHTML={{ __html: t('Filter.Source.CrownOfSorrow') }} />
                   <span dangerouslySetInnerHTML={{ __html: t('Filter.Source.Prestige') }} />
                   <span dangerouslySetInnerHTML={{ __html: t('Filter.Source.Adventure') }} />
                   <span dangerouslySetInnerHTML={{ __html: t('Filter.Source.BlackArmory') }} />